Problem with 24bps files

+++ src/java/org/kc7bfi/jflac/util/WavWriter.java
@@ -158,7 +158,7 @@ public class WavWriter {
                 if (isUnsignedSamples) {
                     for (sample = wideSample = 0; wideSample < wideSamples; wideSample++)
                         for (channel = 0; channel < channels; channel++) {
-                            short val = (short) (channelData[channel].output[wideSample] + 0x800000);
+                            int val = (channelData[channel].output[wideSample] + 0x800000);
                             s8buffer[sample++] = (byte) (val & 0xff);
                             s8buffer[sample++] = (byte) ((val >> 8) & 0xff);
                             s8buffer[sample++] = (byte) ((val >> 16) & 0xff);
@@ -166,7 +166,7 @@ public class WavWriter {
                 } else {
                     for (sample = wideSample = 0; wideSample < wideSamples; wideSample++)
                         for (channel = 0; channel < channels; channel++) {
-                            short val = (short) (channelData[channel].output[wideSample]);
+                            int val = (channelData[channel].output[wideSample]);
                             s8buffer[sample++] = (byte) (val & 0xff);
                             s8buffer[sample++] = (byte) ((val >> 8) & 0xff);
                             s8buffer[sample++] = (byte) ((val >> 16) & 0xff);
